Test support for external Slurm dynamic nodes
Add DevSetting parameter to make the cluster ready for external Slurm
dynamic nodes.

Use ParallelCluster noop prolog if the external Slurm nodes DevSetting is
used.

Share munge key via the Slurm configuration directory on NFS together with
the home of the cluster default user (this last one to be removed in the
future).

Adapt test recipe for connection of external nodes for Slurm dynamic nodes.

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>